pyproject.toml
[project]
name = "bb-browser"
version = "0.1.0"
description = "Web browser for Backboard.io resources"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "backboard-sdk",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
    packages = ["."]